% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barp_partial_dependence.R
\name{barp_partial_dependence}
\alias{barp_partial_dependence}
\title{barp_partial_dependence}
\usage{
barp_partial_dependence(barp.obj, vars = NULL, prop_data = 0.1,
  credible_interval = c(0.025, 0.975), setSeed = NULL)
}
\arguments{
\item{vars}{The variable names or column indices of interest. #' @param vars The variable names or column indices of interest. If a character vector, values at which to evaluate partial dependences based on quantiles c(0.05,seq(.1,.9,by = .1),.95).}

\item{prop_data}{The percentage of the original data to use for estimation. Larger values are more time consuming.}

\item{credible_interval}{The credible interval for the estimates.}

\item{setSeed}{Seed to control random number generation.}

\item{barp}{A \code{\link{barp}} object.}

\item{...}{Additional arguments to be passed to bartMachine.}
}
\value{
Returns an object of class "barpd", containing a list of the following components:
\item{summary}{A \code{data.frame} listing the mean and upper and lower bounds of the predicted outcome at each level of the covariate(s).}
\item{raw}{A \code{data.frame} where each row is a posterior draw and each column is a level for the covariate(s).}
}
\description{
This function calculates the partial dependence for selected covariates based on a \code{BARP} object. The user can specify up to three covariates, resulting in a three-way interaction.
}
\examples{
data("gaymar")
barp.obj <- barp(y = "supp_gaymar",
             x = c("pvote","religcon","age","educ","gXr","stateid","region"),
             dat = svy,
             census = census06,
             geo.unit = "stateid",
             proportion = "n")
bpd <- barp_partial_dependence(barp = barp.obj,
                               vars = list(region = 1:4,educ = 1:4),
                               prop_data = .2,
                               credible_interval = c(0.025,0.975),
                               setSeed = 1021)
}
\seealso{
\code{\link{barp}} which generates the prerequisite \code{barp} object.
}
\keyword{BARP}
\keyword{MRP}
\keyword{Mister}
\keyword{P}
\keyword{multilevel}
\keyword{poststratification}
\keyword{regression}
