% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barp_prognostic_covs.R
\name{barp_prognostic_covs}
\alias{barp_prognostic_covs}
\title{barp_prognostic_covs}
\usage{
barp_prognostic_covs(
  barp.obj,
  interactions = F,
  perm_test = T,
  num_reps = 10,
  num_trees = 20,
  num_permute = 10,
  type = "splits",
  setSeed = NULL,
  ...
)
}
\arguments{
\item{interactions}{A logical statement for whether to evaluate individual variables or their pairwise interactions. Defaults to \code{FALSE}.}

\item{perm_test}{A logical statement for whether to evaluate covariate significance via permutation tests. Defaults to \code{FALSE}.}

\item{num_reps}{The number of reps used to estimate the average inclusion proportions and standard deviation.}

\item{num_trees}{The number of trees to be used.}

\item{num_permute}{The number of permutation simulations.}

\item{type}{The context in which to evaluate the proportion of variable inclusion, either in terms of \code{trees} or in terms of \code{splits}.}

\item{setSeed}{Seed to control random number generation.}

\item{BARP}{A \code{BARP} object.}
}
\value{
Returns an object of class "barpcov", containing a list of the following components:
\item{covariate_importance}{A \code{matrix} containing the raw estimated inclusion proportions for the covariates.}
\item{type}{An indicator for which \code{type} was used to estimate variable importance.}
\item{permutation_test}{A \code{matrix} containing the raw estimated inclusion proportions for the covariates estimated under the permuted null relationship. Only included if \code{perm_test = TRUE}.}
\item{p_vals}{A \code{vector} containing the p-values for each variable from the permutation test. These specifically correspond to the percentabe of the null proportions that are less than or equal to the observed average proportion. Only included if \code{perm_test = TRUE}.}
}
\description{
This function extracts the relative prognostic power of the covariates used in the BARP model. The user can choose to evaluate the statistical significance of these metrics by setting perm_test to TRUE. Doing so re-esetimates variable inclusion proportions (in either trees or splits) after randomly permuting the outcome variable. These permuted inclusion proportions are used as a null distribution against which the average observed proportions are compared.
}
\examples{
data("gaymar")
barp.obj <- barp(y = "supp_gaymar",
             x = c("pvote","religcon","age","educ","gXr","stateid","region"),
             dat = svy,
             census = census06,
             geo.unit = "stateid",
             proportion = "n")
barpcov <- barp_prognostic_covs(barp.obj,
                                interactions = F,
                                perm_test = T,
                                num_reps = 30,
                                num_trees = 20,
                                num_permute = 30,
                                type = 'splits')
}
\seealso{
\code{\link{BARP}} which generates the prerequisite \code{BARP} object.
}
\keyword{BARP}
\keyword{MRP}
\keyword{Mister}
\keyword{P}
\keyword{multilevel}
\keyword{poststratification}
\keyword{regression}
