% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.barp.R
\name{plot.brp}
\alias{plot.brp}
\title{plot.brp}
\usage{
\method{plot}{brp}(barp, evaluate_model = F)
}
\arguments{
\item{barp}{A \code{\link{barp}} object.}

\item{evaluate_model}{Plot model convergence diagnostics? If left to the default \code{FALSE}, return a plot of predictions and credible intervals.}
}
\description{
This function plots the predictions by geographic unit \code{geo.unit}.
}
\examples{
data("gaymar")
barp.obj <- barp(y = "supp_gaymar",
             x = c("pvote","religcon","age","educ","gXr","stateid","region"),
             dat = svy,
             census = census06,
             geo.unit = "stateid",
             proportion = "n")
plot(barp.obj,
     evaluate_model = F)
}
\seealso{
\code{\link{barp}} which creates the prerequisite \code{barp} object.
}
\keyword{BARP}
\keyword{MRP}
\keyword{Mister}
\keyword{P}
\keyword{multilevel}
\keyword{poststratification}
\keyword{regression}
