% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barp_plot_progcov.R
\name{plot.barpcov}
\alias{plot.barpcov}
\title{plot.barpcov}
\usage{
\method{plot}{barpcov}(barpcov, var_names = NULL, sig_level = NULL)
}
\arguments{
\item{barpcov}{A \code{BARP} covariate importance object.}

\item{var_names}{A vector of variable names. If \code{NULL} (the default) the variable names are taken from the \code{BARP} covariate importance object.}

\item{sig_level}{The significance level at which to evaluate covariate significance. Only applicable for BARP covariate importance objects that included a permutation test. Defaults to \code{0.05}.}
}
\description{
This function plots the covariate importance.
}
\examples{
data("gaymar")
BARP <- barp(y = "supp_gaymar",
             x = c("pvote","religcon","age","educ","gXr","stateid","region"),
             dat = svy,
             census = census06,
             geo.unit = "stateid",
             proportion = "n")
barpcov <- barp_prognostic_covs(BARP,
                                interactions = F,
                                perm_test = T,
                                num_reps = 30,
                                num_trees = 20,
                                num_permute = 30,
                                type = 'splits')
plot.barpcov(barpcov,
             var_names = c("Presidential Vote","Religious-Conservative",
                           "Age","Education","Gender X Race","State","Region"),
             sig_level = 0.05)
}
\seealso{
\code{\link{barp_prognostic_covs}} which creates the prerequisite \code{barpcov} object.
}
\keyword{BARP}
\keyword{MRP}
\keyword{Mister}
\keyword{P}
\keyword{multilevel}
\keyword{poststratification}
\keyword{regression}
